contextless kernel through context delegation
